refactor: Update Celery configuration to use jobs module
